+++ src/title_screen.cpp
@@ -34,8 +34,8 @@ void UpdateAndDrawTitleScreen(TitleScreen* title) {
     float p = title->time / titleAnimationTime;
 
     p = min(1.0f, p);
-    Vector2 startPos = {windowWidth / 2.0f - title->logo.width / 2.0f, -title->logo.height};
-    Vector2 endPos   = {windowWidth / 2.0f - title->logo.width / 2.0f, windowHeight / 4};
+    Vector2 startPos = {windowWidth / 2.0f - title->logo.width / 2.0f, (float) -title->logo.height};
+    Vector2 endPos   = {windowWidth / 2.0f - title->logo.width / 2.0f, windowHeight / 4.0f};
     
     Vector2 pos = Vector2Lerp(startPos, endPos, p);
